# =act clean up ByteString#drop(...)