--- a/akka-actor/src/main/scala/akka/util/ByteString.scala
+++ b/akka-actor/src/main/scala/akka/util/ByteString.scala
@@ -485,20 +485,20 @@ object ByteString {
       //            however then we're forced to call .toVector which halfs performance
       //            We can work around that, as there's a Scala private method "remainingVector" which is fast,
       //            but let's not go into calling private APIs here just yet.
-      @tailrec def go(head: Int, restToDrop: Int): (Int, Int) = {
-        val bs = bytestrings(head)
-        if (bs.length > restToDrop) (head, restToDrop)
-        else go(head + 1, restToDrop - bs.length)
+      @tailrec def findSplit(fullDrops: Int, remainingToDrop: Int): (Int, Int) = {
+        val bs = bytestrings(fullDrops)
+        if (bs.length > remainingToDrop) (fullDrops, remainingToDrop)
+        else findSplit(fullDrops + 1, remainingToDrop - bs.length)
       }
 
-      val (head, restToDrop) = go(0, n)
+      val (fullDrops, remainingToDrop) = findSplit(0, n)
 
-      if (restToDrop == 0)
-        new ByteStrings(bytestrings.drop(head), length - n)
-      else if (head == bytestrings.length - 1)
-        bytestrings(head).drop(restToDrop)
+      if (remainingToDrop == 0)
+        new ByteStrings(bytestrings.drop(fullDrops), length - n)
+      else if (fullDrops == bytestrings.length - 1)
+        bytestrings(fullDrops).drop(remainingToDrop)
       else
-        new ByteStrings(bytestrings(head).drop1(restToDrop) +: bytestrings.drop(head + 1), length - n)
+        new ByteStrings(bytestrings(fullDrops).drop1(remainingToDrop) +: bytestrings.drop(fullDrops + 1), length - n)
     }
 
     protected def writeReplace(): AnyRef = new SerializationProxy(this)