Current implementation has a good algorithm but seems a little bit complicated.

Clening-up does not suffer the performance (actually seems to have the better performance when dropping(N-1) where N is the length), and is easy to understand almost the same algorithm now.

### Benchmark
I measured the performance using sbt-jmh with this [code](https://gist.github.com/monkey-mas/1ea2f9c3ba5c7fe8fbd16cf1caf30c8f).  Benchmark code was executed with some parameters such as `-t1 -f2 -wi 10 -i 20`.

#### Test environment
```
OS: Ubuntu 14.0.4 (Linux kernel version 3.13.0-95-generic)
CPU: Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz
VM version: JDK 1.8.0_77
```

#### Before
```
Benchmark                         Mode  Cnt           Score          Error  Units
Drop.drop_avg                   thrpt   40      560267.149 ±     4490.709  ops/s
Drop.drop_best                  thrpt   40    11597871.354 ±    70215.754  ops/s
Drop.drop_greater_or_eq_to_len  thrpt   40   924849361.936 ±  4149839.121  ops/s
Drop.drop_negative              thrpt   40  1209035501.266 ± 18455513.298  ops/s
Drop.drop_worst                 thrpt   40      235822.392 ±     6203.663  ops/s
```

#### After
```
Drop.drop_avg                   thrpt   40      563309.938 ±    5294.201  ops/s
Drop.drop_best                  thrpt   40    11020064.756 ±  322643.144  ops/s
Drop.drop_greater_or_eq_to_len  thrpt   40   922781749.091 ± 5329170.729  ops/s
Drop.drop_negative              thrpt   40  1231164383.713 ± 9861870.265  ops/s
Drop.drop_worst                 thrpt   40      317676.017 ±    1439.986  ops/s
```