@ktoso 2016-09-22
Someone else want to sanity check before merge?