@monkey-mas 2016-09-24
Review please :) If the commits are okay, I'll squash them!

- fixed some tests
- fixed private modifier
- renamed methods and variables
- added benchmark